Share more code between local and remote postMessage https://bugs.webkit.org/show_bug.cgi?id=254773
Reviewed by Chris Dumez.

Share the shape all the way down to the IDL.

* Source/WebCore/CMakeLists.txt:
* Source/WebCore/DerivedSources-input.xcfilelist:
* Source/WebCore/DerivedSources-output.xcfilelist:
* Source/WebCore/DerivedSources.make:
* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/DOMWindow.cpp:
(WebCore::DOMWindow::createTargetOriginForPostMessage):
* Source/WebCore/page/DOMWindow.h:
(WebCore::DOMWindow::identifier const):
* Source/WebCore/page/DOMWindow.idl: Copied from Source/WebCore/page/DOMWindow.h.
* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::postMessage):
* Source/WebCore/page/LocalDOMWindow.h:
(WebCore::WindowPostMessageOptions::WindowPostMessageOptions): Deleted.
* Source/WebCore/page/LocalDOMWindow.idl:
* Source/WebCore/page/RemoteDOMWindow.cpp:
(WebCore::RemoteDOMWindow::postMessage):
* Source/WebCore/page/RemoteDOMWindow.h:
* Source/WebCore/page/RemoteDOMWindow.idl:
* Source/WebCore/page/WindowPostMessageOptions.h: Copied from Source/WebCore/page/DOMWindow.h.
(WebCore::WindowPostMessageOptions::WindowPostMessageOptions):
* Source/WebCore/page/WindowPostMessageOptions.idl: Copied from Source/WebCore/page/DOMWindow.h.
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/262437@main
