+++ Source/WebCore/Headers.cmake
@@ -1369,6 +1369,7 @@ set(WebCore_PRIVATE_FRAMEWORK_HEADERS
     page/WheelEventTestMonitor.h
     page/WindowFeatures.h
     page/WindowOrWorkerGlobalScope.h
+    page/WindowPostMessageOptions.h
     page/WorkerClient.h
 
     page/csp/CSPViolationReportBody.h
