+++ Source/WebCore/Sources.txt
@@ -4338,6 +4338,7 @@ JSWebXRWebGLLayer.cpp
 JSWheelEvent.cpp
 JSWindowEventHandlers.cpp
 JSWindowEventHandlers+Gamepad.cpp
+JSWindowPostMessageOptions.cpp
 JSWindowOrWorkerGlobalScope.cpp
 JSWorker.cpp
 JSWorkerGlobalScope.cpp
