+++ Source/WebCore/page/WindowPostMessageOptions.idl
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Apple Inc. All rights reserved.
+ * Copyright (C) 2023 Apple Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -23,46 +23,6 @@
  * THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#pragma once
-
-#include "EventTarget.h"
-#include "GlobalWindowIdentifier.h"
-#include <wtf/HashMap.h>
-#include <wtf/RefCounted.h>
-
-namespace WebCore {
-
-class Document;
-class Frame;
-class SecurityOrigin;
-
-class DOMWindow : public RefCounted<DOMWindow>, public EventTarget {
-    WTF_MAKE_ISO_ALLOCATED(DOMWindow);
-public:
-    virtual ~DOMWindow();
-
-    static HashMap<GlobalWindowIdentifier, DOMWindow*>& allWindows();
-
-    const GlobalWindowIdentifier& identifier() const { return m_identifier; }
-    virtual Frame* frame() const = 0;
-
-    virtual bool isLocalDOMWindow() const = 0;
-    virtual bool isRemoteDOMWindow() const = 0;
-
-    using RefCounted::ref;
-    using RefCounted::deref;
-
-protected:
-    explicit DOMWindow(GlobalWindowIdentifier&&);
-
-    ExceptionOr<RefPtr<SecurityOrigin>> createTargetOriginForPostMessage(const String&, Document&);
-
-    EventTargetInterface eventTargetInterface() const final { return LocalDOMWindowEventTargetInterfaceType; }
-    void refEventTarget() final { ref(); }
-    void derefEventTarget() final { deref(); }
-
-private:
-    GlobalWindowIdentifier m_identifier;
+dictionary WindowPostMessageOptions : StructuredSerializeOptions {
+    USVString targetOrigin = "/";
 };
-
-} // namespace WebCore
